requirements.txt
faster-whisper>=1.0.0
PyYAML>=6.0.1
tqdm>=4.66.0
rich>=13.7.0
pytest>=8.0.0
vosk==0.3.45
soundfile==0.12.1
